+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Fixed
 
+- Missing loot table for some blocks
+
 ## [1.4.0] - 2024-05-10
 
 ### Added